+++ react-components/src/components/Table/model/i-property-grid.ts
@@ -1,0 +1,8 @@
+import { IDictionary } from './dictionary';
+
+export interface IPropertyGrid<T> {
+  value: T;
+  columnName?: string;
+  dict?: IDictionary;
+  onChange?: (value: T) => void;
+}